Feat: Add response methods for 202 Accepted and 503 Service Unavailable, Renamed util and route file from *_util.js to *.util.js and file from *_route.js to *.route.js
