+++ src/routes/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const HelloWorld = require("./hello_world_route.js");
+const HelloWorld = require("./hello_world.route");
 
 router.use(HelloWorld);
