+++ src/utils/http_status.util.js
@@ -2,7 +2,7 @@ const HttpStatus = {
     // Success Responses
     OK: 200,
     CREATED: 201,
-    // ACCEPTED: 202,
+    ACCEPTED: 202,
     NO_CONTENT: 204,
 
     // Client Errors
@@ -15,7 +15,7 @@ const HttpStatus = {
 
     // Server Errors
     INTERNAL_SERVER_ERROR: 500,
-    // SERVICE_UNAVAILABLE: 503,
+    SERVICE_UNAVAILABLE: 503,
 };
 
 module.exports = HttpStatus;
