+++ src/utils/index.js
@@ -1,9 +1,10 @@
-const EncryptionUtil = require("./encryption_util");
-const PasswordUtil = require("./password_util");
-const ResponseUtil = require("./response_util");
+const EncryptionUtil = require("./encryption.util");
+const PasswordUtil = require("./password.util");
+const { HttpStatus, ResponseUtil } = require("./response.util");
 
 module.exports = {
     EncryptionUtil,
+    HttpStatus,
     PasswordUtil,
     ResponseUtil,
 };
